Add aliases to main and renderer
This required a change to typescript, webpack and eslint
configurations. The eslint part is handled using
eslint-import-resolver-typescript. The typescript and webpack parts are
synchronized manually. There's a plugin that handles webpack
configuration by reading tsconfig, but it interfered with the hot reload
alias. So I elected not to use it.