Base useReposittory on Commits class
The Commits class semi-wraps the commits which are now saved as an
array, and adds wrapper for easy access by hash. It also encapsulates
all the logic of turning the raw commit info into the full structure that
includes information for graphs.
isLoading is now part of the useRepository hook to differentiate
between being outside a repository and waiting for it to load.