😱 MAJOR REFACTOR
The main focus was on encapsulating the concept of the repository.
Repository is now a class that contains Commits and Refs, and is part
of the git utils. The useRepository hook was greatly simplified and all
of the code was updated to support the changes.

Other refactors were applied as well, in the following areas:
1. `BranchBadges` exported its logic of combining refs to a utils file
2. `PlacementStrategy` doesn't deal with row ordering anymore, as this
is specified in the `git log` command.